% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{copyMOVESDatabase}
\alias{copyMOVESDatabase}
\title{copyMOVESDatabase}
\usage{
copyMOVESDatabase(dbconn, db_from_name, db_to_name)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{db_from_name}{MySQL county database to be copied as string.}

\item{db_to_name}{MySQL county database to be created as string.}
}
\value{
boolean if the database was successfully copied
}
\description{
Copies an existing MOVES database to a new location.
}
\examples{
copyMOVESDatabase(dbconn, countydb_name, new_countydb_name)
}
