% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{createBatchFile}
\alias{createBatchFile}
\title{createBatchFile}
\usage{
createBatchFile(batchFileLocation, runspecLocations, movesLocation)
}
\arguments{
\item{batchFileLocation}{file path as string as location to save batchfile}

\item{runspecLocations}{vector of file paths as string as location to find runspecs}

\item{movesLocation}{file path as string as location of MOVES executable installation}
}
\value{
file path as string as location to where batchfile was saved
}
\description{
Creates a MOVES excutable batchfile that will run all runspecs provided.
}
\examples{
createBatchFile("C:\\\\Users\\\\LocalUser\\\\MOVESRuns\\\\moves_batchfile.bat",
c("C:\\\\Users\\\\LocalUser\\\\MOVESRuns\\\\moves_runspec_copy.xml"),
"C:\\\\Users\\\\Public\\\\EPA\\\\MOVES\\\\MOVES2014b")
}
