% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{deleteMOVESRun}
\alias{deleteMOVESRun}
\title{deleteMOVESRun}
\usage{
deleteMOVESRun(dbconn, outputdb_name, movesrunid)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{outputdb_name}{MySQL county database to be updated as string.}

\item{movesrunid}{MOVESRunID for rows to be deleted}
}
\value{


TRUE if succesful, FALSE otherwise
}
\description{
deleteMOVESRun
}
\examples{
deleteMOVESRun(dbconn, outputdb_name, 6)
}
