% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{endDBConnection}
\alias{endDBConnection}
\title{endDBConnection}
\usage{
endDBConnection(dbconn)
}
\arguments{
\item{dbconn}{}
}
\value{
BOOLEAN result of \code{dbDisconnect()}
}
\description{
A wrapper for \code{dbDisconnect()}
}
\examples{
endDBConnection(dbconn)
}
