% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{getMOVESBaseTable}
\alias{getMOVESBaseTable}
\title{getMOVESBaseTable}
\usage{
getMOVESInputTable(dbconn, movesdb_name, table_name)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{movesdb_name}{MySQL default database to be updated as string}

\item{table_name}{MySQL table to be updated as string}
}
\value{
Either a dataframe with the result from \code{RMySQL::dbSendQuery()} or FALSE
}
\description{
Gets the results of a base table in a MOVES database.
}
\examples{
getMOVESBaseTable(dbconn, movesdb_name, "sourceusetype")
}
