% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{getMOVESOutputTable}
\alias{getMOVESOutputTable}
\title{getMOVESOutputTable}
\usage{
getMOVESOutputTable(
  dbconn,
  movesdb_name,
  outputdb_name,
  table_name,
  get_query_string = FALSE,
  moves_run_id = FALSE
)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{movesdb_name}{MySQL default database to be updated as string}

\item{outputdb_name}{MySQL output database to be updated as string}

\item{table_name}{MySQL table to be updated as string}

\item{get_query_string}{TRUE/FALSE if TRUE the query string is returned instead of the results, defaults to FALSE#'}
}
\value{
Either a dataframe with the result from \code{RMySQL::dbSendQuery()} or FALSE
}
\description{
Gets the results of a table in a MOVES database and all associated lookup tables.
}
\examples{
getMOVESOutputTable(dbconn, movesdb_name, outputdb_name, "movesoutput")
getMOVESOutputTable(dbconn, movesdb_name, outputdb_name, "rateperdistance", get_query_string = TRUE)
getMOVESOutputTable(dbconn, movesdb_name, outputdb_name, "movesoutput", moves_run_id = 3)
}
