% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{getMOVESTables}
\alias{getMOVESTables}
\title{getMOVESTableS}
\usage{
getMOVESTables(dbconn, db_name)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{movesdb_name}{MySQL default database to be updated as string}
}
\value{
Either a dataframe with the result from \code{dbSendQuery()} or FALSE
}
\description{
Gets the results of a table in a MOVES database.
}
\examples{
getMOVESTableS(dbconn, db_name)
}
