% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{getRunspecAttr}
\alias{getRunspecAttr}
\title{getRunspecAttr}
\usage{
getRunspecAttr(runspec, xpathsLocation, attribute)
}
\arguments{
\item{runspec}{\code{XML::xmlParse()} object}

\item{xpathsLocation}{string in the format of an XPath (https://way2tutorial.com/xml/xpath-expression.php)}

\item{attribute}{the name of the attribute being found}
}
\value{
\code{XML::xpathSApply()} of attribute from runspec
}
\description{
Gets an attibute in a runspec \code{XML::xmlParse()} object.
}
\examples{
getRunspecAttr(rs, "//scaleinputdatabase", "databasename")
}
