% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{getRunspecValue}
\alias{getRunspecValue}
\title{getRunspecValue}
\usage{
getRunspecValue(runspec, xpathsLocation)
}
\arguments{
\item{runspec}{\code{XML::xmlParse()} object}

\item{xpathsLocation}{string in the format of an XPath (https://way2tutorial.com/xml/xpath-expression.php)}
}
\value{
\code{XML::xNodeSet()} of the node from runspec
}
\description{
Gets an attibute in a runspec \code{XML::xmlParse()} object.
}
\examples{
getRunspecValue(rs, "//scaleinputdatabase")
}
