% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{noWeekDays}
\alias{noWeekDays}
\title{noWeekDays}
\usage{
noWeekDays(sd, ed, weekend = c("Saturday", "Sunday"))
}
\arguments{
\item{sd}{start date (as.Date) of timespan}

\item{ed}{end date (as.Date) of timespan}

\item{weekend}{vector of text names of weekend days, defaults to c("Saturday", "Sunday")}
}
\value{
noWeekDays integer
}
\description{
Returns integer sum of number of weekdays in timespan from a to b
}
\examples{
noWeekDays(as.Date("1969-01-01", "2012-12-21"))
noWeekDays(as.Date("2016-01-21", "2020-01-21", weekend=c("Sunday")))
}
