% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{processGetQuery}
\alias{processGetQuery}
\title{processGetQuery}
\usage{
processGetQuery(dbconn, query, get_query_string = FALSE)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{query}{string with they MySQL Query}

\item{get_query_string}{TRUE/FALSE if TRUE the query string is returned instead of the results, defaults to FALSE}
}
\value{
Either a dataframe with the result from \code{dbSendQuery()} or FALSE
}
\description{
processGetQuery
}
\examples{
processGetQuery(dbconn, "select * from movesrun")
processGetQuery(dbconn, "select * from movesrun", TRUE)
}
