% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{readRunspec}
\alias{readRunspec}
\title{readRunspec}
\usage{
readRunspec(runspecLocation)
}
\arguments{
\item{runspecLocation}{file path as string}
}
\value{
\code{XML::xmlParse()} of runSpec
}
\description{
Reads in a runspec at the location provided into an \code{XML::xmlParse()} object.
}
\examples{
readRunspec("C:\\\\Users\\\\LocalUser\\\\MOVESRuns\\\\moves_runspec.xml")
}
