% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{renumberMOVESRun}
\alias{renumberMOVESRun}
\title{renumberMOVESRun}
\usage{
renumberMOVESRun(dbconn, outputdb_name, oldmovesrunid, newmovesrunid)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{outputdb_name}{MySQL county database to be updated as string.}

\item{oldmovesrunid}{MOVESRunID for rows to be moved}

\item{newmovesrunid}{MOVESRunID for where rows are to be moved}
}
\value{
TRUE if succesful, FALSE otherwise
}
\description{
renumberMOVESRun
}
\examples{
renumberMOVESRun(dbconn, outputdb_name, 6, 4)
}
