% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{replaceMOVESTable}
\alias{replaceMOVESTable}
\title{replaceMOVESTable}
\usage{
replaceMOVESTable(dbconn, db_name, table_name, data)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{db_name}{MySQL county database to be updated as string.}

\item{table_name}{MySQL table to be updated as string.}

\item{data}{Data from a MySQL query of a MOVES table.}
}
\value{
boolean if the database was successfully replaced
}
\description{

}
\examples{
replaceMOVESTable(dbconn, new_countydb_name, "avgspeeddistribution", data)
}
