% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{runMOVES}
\alias{runMOVES}
\title{runMOVES}
\usage{
runMOVES(batchFileLocation)
}
\arguments{
\item{batchFileLocation}{}
}
\value{
results of command line MOVES run.
}
\description{
Runs command line MOVES using the batchfile supplied.
}
\examples{
runMOVES("C:\\\\Users\\\\LocalUser\\\\MOVESRuns\\\\moves_batchfile.bat")
}
