% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{setRunspecAttr}
\alias{setRunspecAttr}
\title{setRunspecAttr}
\usage{
setRunspecAttr(runspec, xpathsLocation, list)
}
\arguments{
\item{runspec}{\code{XML::xmlParse()} object}

\item{xpathsLocation}{string in the format of an XPath (https://way2tutorial.com/xml/xpath-expression.php)}

\item{list}{vector containing attributes and the values the attributes should be set to}
}
\value{

}
\description{
Sets an attibute in a runspec \code{XML::xmlParse()} object.
}
\examples{
setRunspecAttr(rs, "//outputdatabase", c(databasename =  new_outputdb_name))
}
