% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{setRunspecValue}
\alias{setRunspecValue}
\title{setRunspecValue}
\usage{
setRunspecValue(runspec, xpathsLocation, value, cdata = FALSE)
}
\arguments{
\item{runspec}{\code{XML::xmlParse()} object}

\item{xpathsLocation}{string in the format of an XPath (https://way2tutorial.com/xml/xpath-expression.php)}

\item{value}{string to set the value in the selected node to
#'}
}
\value{
\code{XML::xNodeSet()} of the node from runspec
}
\description{
Sets an attibute in a runspec \code{XML::xmlParse()} object.
}
\examples{
setRunspecValue(rs, "//description", "MOVES Test Run 2")
}
