% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{deleteMOVESRun}
\alias{deleteMOVESRun}
\title{deleteMOVESRun}
\usage{
deleteMOVESRun(dbconn, movesdb_name, outputdb_name, movesrunid)
}
\arguments{
\item{movesrunid}{}
}
\value{

}
\description{
deleteMOVESRun
}
