% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{getMOVESBaseTable}
\alias{getMOVESBaseTable}
\title{getMOVESBaseTable}
\usage{
getMOVESBaseTable(dbconn, movesdb_name, table_name, get_query_string = FALSE)
}
\arguments{
\item{dbconn}{MySQL db connection}

\item{movesdb_name}{MySQL default database to be updated as string}

\item{table_name}{MySQL table to be updated as string}

\item{get_query_string}{TRUE/FALSE if TRUE the query string is returned instead of the results, defaults to FALSE}
}
\value{
Either a dataframe with the result from \code{RMySQL::dbSendQuery()} or FALSE
}
\description{
Gets the results of a table in a MOVES database.
}
\examples{
getMOVESBaseTable(dbconn, movesdb_name, "sourceusetype")
}
