% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{makeDBConnection}
\alias{makeDBConnection}
\title{makeDBConnection}
\usage{
makeDBConnection(
  user,
  password,
  host = "localhost",
  port = 3306,
  dbapi = "MariaDB"
)
}
\arguments{
\item{user}{mysql user name}

\item{password}{mysql password}

\item{host}{mysql host, default to localhost}
}
\value{
\code{dbConnect()}
}
\description{
A wrapper for \code{dbConnect()}
}
\examples{
makeDBConnection("mysql_user", "password12345")
makeDBConnection("mysql_user", "password12345", host="168.1.1.1")
}
