% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4moves.R
\name{renumberMOVESRun}
\alias{renumberMOVESRun}
\title{renumberMOVESRun}
\usage{
renumberMOVESRun(
  dbconn,
  movesdb_name,
  outputdb_name,
  oldmovesrunid,
  newmovesrunid
)
}
\arguments{
\item{newmovesrunid}{}
}
\value{

}
\description{
renumberMOVESRun
}
